rename annotation file `*_test.dart` to `*_test_annotation.dart`
